// test/integration/test/ports/_fixtures/blog-update/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
type DefaultLiteralValue<CodecId extends string, _Encoded> = CodecId extends keyof CodecTypes
  ? CodecTypes[CodecId]['output']
  : _Encoded;

export type FieldOutputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly optional: CodecTypes['pg/text@1']['output'] | null;
      readonly authorId: CodecTypes['pg/text@1']['output'] | null;
      readonly lastReviewedAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly lastPublishedAt: CodecTypes['pg/timestamptz@1']['output'] | null;
    };
    readonly Profile: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly notrequired: CodecTypes['pg/text@1']['output'] | null;
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly goesToBedAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly goesToOfficeAt: CodecTypes['pg/timestamptz@1']['output'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly wakesUpAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly lastLoginAt: CodecTypes['pg/timestamptz@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly optional: CodecTypes['pg/text@1']['input'] | null;
      readonly authorId: CodecTypes['pg/text@1']['input'] | null;
      readonly lastReviewedAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly lastPublishedAt: CodecTypes['pg/timestamptz@1']['input'] | null;
    };
    readonly Profile: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly notrequired: CodecTypes['pg/text@1']['input'] | null;
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly goesToBedAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly goesToOfficeAt: CodecTypes['pg/timestamptz@1']['input'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly wakesUpAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly lastLoginAt: CodecTypes['pg/timestamptz@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Post: {
      readonly author: CodecTypes['pg/text@1']['output'] | null;
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly lastPublishedAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly lastReviewedAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly optional: CodecTypes['pg/text@1']['output'] | null;
      readonly published: CodecTypes['pg/bool@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['output'];
    };
    readonly Profile: {
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly goesToBedAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly goesToOfficeAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly notrequired: CodecTypes['pg/text@1']['output'] | null;
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly lastLoginAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly wakesUpAt: CodecTypes['pg/timestamptz@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Post: {
      readonly author: CodecTypes['pg/text@1']['input'] | null;
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly lastPublishedAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly lastReviewedAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly optional: CodecTypes['pg/text@1']['input'] | null;
      readonly published: CodecTypes['pg/bool@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['input'];
    };
    readonly Profile: {
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly goesToBedAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly goesToOfficeAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly notrequired: CodecTypes['pg/text@1']['input'] | null;
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly lastLoginAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly wakesUpAt: CodecTypes['pg/timestamptz@1']['input'] | null;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;